feat: Add support for Kubernetes contexts
This change introduces the ability to specify a Kubernetes context for a podbook configuration.

A new `podbook-contexts` variable can be used to map a short identifier to a full Kubernetes context name. Individual entries in `podbook-configurations` can then specify a `:context-id` to use the corresponding context.

The `podbook` object is updated to store this context information, and a new helper function `podbook--kubectl` is added to construct `kubectl` commands with the correct `--context` flag.